+++ scripts/plot_predictions.py
@@ -29,8 +29,8 @@ def parse_args():
 
 def _timestamps_to_float(col: pd.Series) -> np.ndarray:
     if np.issubdtype(col.dtype, np.datetime64):
-        return (col.view("int64") / 1e9).astype("float32")
-    return col.astype("float32").to_numpy()
+        return (col.view("int64") / 1e9).astype("float64")
+    return col.astype("float64").to_numpy()
 
 def main():
     args = parse_args()
