+++ src/ts_transformer/plot_test_predictions.py
@@ -32,9 +32,9 @@ def _timestamps_to_float(col: pd.Series) -> np.ndarray:
     - numérico -> float directamente
     """
     if np.issubdtype(col.dtype, np.datetime64):
-        return (col.view("int64") / 1e9).astype("float32")
+        return (col.view("int64") / 1e9).astype("float64")
     else:
-        return col.astype("float32").to_numpy()
+        return col.astype("float64").to_numpy()
 
 
 def parse_args() -> argparse.Namespace:
